Fix parsertests so that they work even when server-side thumbnailing isn't configured.

+++ includes/MessageCache.php
@@ -399,16 +399,18 @@ class MessageCache {
 
 		// Also delete cached sidebar... just in case it is affected
 		global $parserMemc;
-		$codes = array( $code );
-		if ( $code === 'en'  ) {
-			// Delete all sidebars, like for example on action=purge on the
-			// sidebar messages
-			$codes = array_keys( Language::getLanguageNames() );
-		}
+		if(is_object($parserMemc)) {
+			$codes = array( $code );
+			if ( $code === 'en'  ) {
+				// Delete all sidebars, like for example on action=purge on the
+				// sidebar messages
+				$codes = array_keys( Language::getLanguageNames() );
+			}
 
-		foreach ( $codes as $code ) {
-			$sidebarKey = wfMemcKey( 'sidebar', $code );
-			$parserMemc->delete( $sidebarKey );
+			foreach ( $codes as $code ) {
+				$sidebarKey = wfMemcKey( 'sidebar', $code );
+				$parserMemc->delete( $sidebarKey );
+			}
 		}
 
 		wfRunHooks( "MessageCacheReplace", array( $title, $text ) );
